+++ CHANGELOG.md
@@ -16,6 +16,7 @@ References and Resolving:
 1.  Resolve `ModuleImport` declarations that don't specify `at` paths.
 1.  Resolve `NamespaceDecl` declarations.
 1.  Resolve `DefaultNamespaceDecl` declarations.
+1.  Resolve `DirAttributeList` attributes.
 
 ## 1.2 - 2018-08-27